requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
PyYAML>=6.0.1
openpyxl>=3.1.2
pytest>=8.0.0
